Web: Affiliate Tracking
Added relevant columns for affiliate tracking to the users table.

Added script during the creation of user to generate a unique random 10 character string as the tracking id (str_random(10)).

Added a [CheckReferral] middleware to scan incoming queries for the variable [ref] & attach cookies containing the referral's tracking code on to request.

Added script during the user's creation to check for referral cookies & attach on to the [referred_by] column.